fixed id not found issue

+++ view.php
@@ -28,14 +28,18 @@
     <tbody>
   <?php
     include './connection.php';
+    // print $_GET['id'];
+    if (isset($_GET['id'])) {
+      $id = $_GET['id'];
+
+      $delete_query = "DELETE FROM `user_details` WHERE `user_details`.`id` = '$id'";
+      $res = $con->query($delete_query);
+      if ($res) { print 'pass1'; }
+    }
+
     $read = "SELECT * FROM `user_details`";
     $result = $con->query($read);
-
-    $id = $_GET['id'];
-
-    $delete_query = "DELETE FROM `user_details` WHERE `user_details`.`id` = '$id'";
-    $res = $con->query($delete_query);
-    if ($res) { print 'pass1'; }
+    
     // $result = mysqli_query($con, $read);
     if ($result -> num_rows > 0) {
       // print 'pass';
